# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leave Management API",
        "description": "A comprehensive Leave Management API with role-based access control for managing employee leave requests, approvals, and leave balance tracking.",
        "contact": {
            "email": "support@leavemanagement.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/leave-requests/{id}": {
            "get": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Get a specific leave request",
                "description": "Retrieve details of a specific leave request. General users can only view their own requests.",
                "operationId": "25fb1eb2f397b89937e6e5cae3e0b3c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leave request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leave request retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "leave_type": {
                                                    "type": "string",
                                                    "example": "full_day"
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "2026-01-20"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "example": "2026-01-20"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "approved"
                                                },
                                                "days_count": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Family emergency"
                                                },
                                                "user": {
                                                    "type": "object"
                                                },
                                                "approver": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to view this leave request"
                    },
                    "404": {
                        "description": "Leave request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/leave-requests/{id}/approve": {
            "post": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Approve a leave request",
                "description": "Approve a pending leave request. HR can approve General user requests. Admin can approve HR requests. Admins cannot approve General user requests directly.",
                "operationId": "198e9fcdf7865f0509180e7ca3a38c78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leave request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leave request approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Leave request approved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Leave already processed or invalid approval chain"
                    },
                    "403": {
                        "description": "Unauthorized - Insufficient permissions"
                    },
                    "404": {
                        "description": "Leave request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/leave-requests/{id}/reject": {
            "post": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Reject a leave request",
                "description": "Reject a pending leave request with a reason. Same approval hierarchy applies as approval.",
                "operationId": "b8a4c34ed3bcf8cb44c735415c0b9eec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leave request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rejection_reason"
                                ],
                                "properties": {
                                    "rejection_reason": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "minLength": 10,
                                        "example": "Insufficient staffing during requested period"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Leave request rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Leave request rejected"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Leave already processed"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Leave request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}